requirements.txt
aiohttp==3.9.1
aiosignal==1.3.1
altgraph==0.17.4
arrow==0.13.1
asgiref==3.7.2
asyncio==3.4.3
attrs==23.1.0
beautifulsoup4==4.12.2
bs4==0.0.1
cachetools==5.3.1
certifi==2023.5.7
cffi==1.15.1
chardet==3.0.4
charset-normalizer==3.1.0
cloudscraper==1.2.71
colorama==0.4.6
comtypes==1.2.1
discord.py==2.3.2
Django==4.2.7
docopt==0.6.2
et-xmlfile==1.1.0
exceptiongroup==1.1.3
frozenlist==1.4.0
google==3.0.0
google-api-core==2.11.1
google-api-python-client==2.99.0
google-auth==2.23.0
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.1.0
googleapis-common-protos==1.60.0
gspread==5.11.2
h11==0.14.0
httplib2==0.22.0
idna==2.8
image==1.5.33
lxml==5.1.0
multidict==6.0.4
mysql-connector==2.2.9
numpy==1.25.2
oauth2client==4.1.3
oauthlib==3.2.2
openpyxl==3.1.2
outcome==1.2.0
packaging==23.2
pandas==2.1.1
pefile==2023.2.7
Pillow==10.1.0
pipreqs==0.4.13
protobuf==4.23.3
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
PyGetWindow==0.0.9
pyinstaller==6.3.0
pyinstaller-hooks-contrib==2023.12
pyparsing==3.1.1
PyQt5==5.15.9
PyQt5-Qt5==5.15.2
PyQt5-sip==12.12.1
PyRect==0.2.0
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2023.3.post1
pywin32==306
pywin32-ctypes==0.2.2
pywinauto==0.6.8
requests==2.21.0
requests-oauthlib==1.3.1
requests-toolbelt==1.0.0
rsa==4.9
schedule==1.2.1
selenium==4.12.0
six==1.16.0
slackclient==2.9.4
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.5
sqlparse==0.4.4
suds-community==1.1.2
trio==0.22.2
trio-websocket==0.10.4
tzdata==2023.3
undetected-chromedriver==3.5.3
uritemplate==4.1.1
urllib3==1.24.3
websocket-client==1.6.1
websockets==11.0.3
wsproto==1.2.0
XlsxWriter==3.1.6
yarg==0.1.9
yarl==1.9.3
